refactor(seedFile): setup seed script for create game test, move test files
